functions/requirements.txt
firebase_functions~=0.1.0
numpy 
tenacity
python-dateutil

google-cloud-firestore